improvement: restructuring card matkul studi-ku
