+++ src/modules/MyStudy/studiku-dash.tsx
@@ -2,9 +2,8 @@ import Image from "next/image";
 import Dummy from "@/assets/myStudy/Dummy.svg";
 import { FC, Fragment, ReactElement } from "react";
 import Card from "@/components/Common/Card";
-import CardStudy from "@/components/MyStudy/Card";
+import CardStudy from "./CardMatkul/Card";
 import { useDataCard } from "@/hooks/MyStudy/useDataCard";
-import { useGetAllSubject } from "@/hooks/MyStudy/useSubjectCard";
 
 const MyStudy: FC = (): ReactElement => {
   //api
