+++ src/stores/MyStudy/type.d.ts
@@ -5,16 +5,6 @@ export type Data = {
   warna: string;
 };
 
-export type OptionSubject = {
-  src: StaticImageData;
-  namaDosen: string;
-  pertemuan: number;
-  kodematkul: number;
-  tipematkul: number;
-  sks: number;
-  slug: string;
-};
-
 export type Question = {
   id: number;
   question: string;
